Split out `user_interface.h` into `esp_system.h` and `esp_wifi.h`
