--- Sming/Arch/Esp8266/Components/esp_no_wifi/include/esp_wifi.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#include <user_interface.h>
-
-#ifdef __cplusplus
-}
-#endif
